pyproject.toml
[project]
name = "hoopla"
version = "0.1.0"
description = "A movie search engine using BM25 ranking and semantic search algorithms"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.48.0",
    "nltk==3.9.1",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
